check_ip/requirements.txt
beautifulsoup4
lxml
requests==2.24.0
PySocks
